1，dag修改到airflow web不再报错。2，优化airflow启动脚本。3，优化所有deploy脚本，增加执行选项。

+++ env/cpscripts/deploy_master01_dags.sh
@@ -1,29 +1,40 @@
 #如果是项目定制开发的dag部分有修改
 #把k8sdeploy.sh打包生成k8sdeploy.tar.gz上传到master01，解压到home目录
+#!/usr/bin/env bash
 
 :<<EOF
-cd ~/kube-airflow
-
-cp ~/k8sdeploy_dir/dags/* dags/
-cp ~/k8sdeploy_dir/requirements.txt requirements/dags.txt
-
-kubectl delete -f airflow.all.yaml -n fl
-kubectl create -f airflow.all.yaml -n fl
+#example:
+deploy_master01_dags.sh
+deploy_master01_dags.sh -rb
 EOF
 
+rb="no"
+for var in "$*"
+do
+  if [ "$var" == "-rb" ]; then
+    rb="yes"
+  fi
+done
+
+set -x
+
 rm -rf ${HOME}/nfsmnt/dags
-cp -rf ~/k8sdeploy_dir/dags ${HOME}/nfsmnt/
+cp -rvf ~/k8sdeploy_dir/dags ${HOME}/nfsmnt/
 
 #如果dag脚本的python第三方依赖项没有改变，可以跳过image building，只重启airflow
-docker images|grep "<none>"|awk '{print $3}'|xargs docker rmi -f
 
-docker images|grep airflow
-docker images|grep airflow|awk '{print $3}'|xargs docker rmi -f
-sudo ansible slavek8s -m shell -a"docker images|grep airflow|awk '{print \$3}'|xargs docker rmi -f"
-docker images|grep airflow
+if [ "$rb" == "yes" ]; then
+  cd ~/myairflow
 
-docker build -t master01:30500/bronzels/airflow:1.10.10-python3.6 ./
-docker push master01:30500/bronzels/airflow:1.10.10-python3.6
+  docker images|grep "<none>"|awk '{print $3}'|xargs docker rmi -f
 
+  docker images|grep airflow
+  docker images|grep airflow|awk '{print $3}'|xargs docker rmi -f
+  sudo ansible slavek8s -m shell -a"docker images|grep airflow|awk '{print \$3}'|xargs docker rmi -f"
+  docker images|grep airflow
+
+  docker build -t master01:30500/bronzels/airflow:1.10.10-python3.6 ./
+  docker push master01:30500/bronzels/airflow:1.10.10-python3.6
+fi
 
 ~/scripts/myairflow-cp-op.sh restart
